move git code to other file
