+++ core/etl_utils.py
@@ -33,6 +33,7 @@ from pprint import pformat
 import smtplib, ssl
 from pyspark.sql.window import Window
 from pyspark.sql import functions as F
+from core.git_utils import Git_Config_Manager
 import core.logger as log
 logger = log.setup_logging('Job')
 
@@ -60,7 +61,7 @@ class ETL_Base(object):
         self.loaded_inputs = loaded_inputs
         self.jargs = self.set_jargs(pre_jargs, loaded_inputs) if not jargs else jargs
         if self.jargs.manage_git_info:
-            git_yml = Git_Config_Manager().get_config()
+            git_yml = Git_Config_Manager().get_config(LOCAL_APP_FOLDER)
             [git_yml.pop(key, None) for key in ('diffs', 'diffs_yaetos')]
             logger.info('Git info {}'.format(git_yml))
 
@@ -421,59 +422,6 @@ class Meta_Builder():  # TODO: rename to "schemas" here and below
         os.makedirs(os.path.dirname(fname), exist_ok=True)
         with open(fname, 'w') as file:
             ignored = yaml.dump(self.yml, file)
-
-class Git_Config_Manager():
-
-    FNAME = 'conf/git_config.yml'
-
-    def get_config(self):
-        if self.is_git_controlled():
-            config = self.get_config_from_git()
-            self.save_yaml(config)
-        else:
-            config = self.get_config_from_file()
-        return config
-
-    def get_config_from_git(self):
-        import subprocess
-        branch = subprocess.check_output(["git", "describe", '--all']).strip().decode('ascii')  # to get if dirty, add '--dirty'.
-        last_commit = subprocess.check_output(['git', 'rev-parse', 'HEAD']).strip().decode('ascii')
-        diffs = subprocess.check_output(['git', 'diff', 'HEAD']).strip().decode('ascii')
-        is_dirty = True if diffs else False
-        branch_yaetos = subprocess.check_output(['git', 'describe', '--all'], cwd=LOCAL_APP_FOLDER).strip().decode('ascii')
-        last_commit_yaetos = subprocess.check_output(['git', 'rev-parse', 'HEAD'], cwd=LOCAL_APP_FOLDER).strip().decode('ascii')
-        diffs_yaetos = subprocess.check_output(['git', 'diff', 'HEAD'], cwd=LOCAL_APP_FOLDER).strip().decode('ascii')
-        is_dirty_yaetos = True if diffs_yaetos else False
-
-        config = {'branch':branch,
-                  'last_commit':last_commit,
-                  'diffs':diffs,
-                  'is_dirty':is_dirty,
-                  'branch_yaetos':branch_yaetos,
-                  'last_commit_yaetos':last_commit_yaetos,
-                  'diffs_yaetos':diffs_yaetos,
-                  'is_dirty_yaetos':is_dirty_yaetos
-                  }
-        return config
-
-    def is_git_controlled(self):
-        import subprocess
-        out = os.system('git rev-parse')  # not using subprocess.check_output() to avoid crash if it fails.
-        if out == 0:
-            return True
-        else:
-            return False  # will send "fatal: not a git repository" to stderr
-
-    def save_yaml(self, config):
-        os.makedirs(os.path.dirname(self.FNAME), exist_ok=True)
-        with open(self.FNAME, 'w') as file:
-            documents = yaml.dump(config, file)
-
-    def get_config_from_file():
-        if os.path.isfile(self.FNAME):
-            return Job_Yml_Parser.load_meta(self.FNAME)
-        else:
-            return False
 
 
 class Job_Yml_Parser():
